update star count
